fix: update WorkloadUtilsTest revision hash for stretch cluster changes
The pod revision hash changed from 6458a317 to 070283ab due to stretch
cluster code modifications that affect pod generation. The revision is a
SHA1 hash of the entire pod JSON, so any changes to the pod structure
will result in a different hash.

This is expected behavior and the test needs to be updated to reflect
the new hash value.

Signed-off-by: Aswin A <aswin6303@gmail.com>